handle exp-dir path is not a directory

+++ runners.py
@@ -59,6 +59,8 @@ class BaseRunner:
         self.experiment_data_directory = Path(experiment_data_directory) / (
             suite_name + "_" + data_suffix
         )
+        if self.experiment_data_directory.exists() and not self.experiment_data_directory.is_dir():
+            raise RuntimeError(f"Path exists but is not a directory: {self.experiment_data_directory}")
         if fresh:
             if self.experiment_data_directory.exists():
                 shutil.rmtree(self.experiment_data_directory)
